Memory fixes

+++ optimizers.py
@@ -217,7 +217,7 @@ def v_optimize_couple(money,sgrid,umult,EV,sigma,beta,ls,us,use_cp=ucp):
     V_all = V_opt_arr
     
     
-    return ret(V), ret(c), ret(s), ret(i_opt), ret(i_ls), ret(V_all)
+    return ret(V), ret(c), ret(s), ret(i_opt), ret(i_ls), ret(V_all).astype(np.float32)
